// frontend/src/lib/generated/api/models/DTOClaseInscrita.ts
export interface DTOClaseInscrita {
    /**
     * 
     * @type {number}
     * @memberof DTOClaseInscrita
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof DTOClaseInscrita
     */
    titulo?: string;
    /**
     * 
     * @type {string}
     * @memberof DTOClaseInscrita
     */
    descripcion?: string;
    /**
     * 
     * @type {number}
     * @memberof DTOClaseInscrita
     */
    precio?: number;
    /**
     * 
     * @type {string}
     * @memberof DTOClaseInscrita
     */
    presencialidad?: DTOClaseInscritaPresencialidadEnum;
    /**
     * 
     * @type {string}
     * @memberof DTOClaseInscrita
     */
    imagenPortada?: string;
    /**
     * 
     * @type {string}
     * @memberof DTOClaseInscrita
     */
    nivel?: DTOClaseInscritaNivelEnum;
    /**
     * 
     * @type {Array<string>}
     * @memberof DTOClaseInscrita
     */
    alumnosId?: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof DTOClaseInscrita
     */
    profesoresId?: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof DTOClaseInscrita
     */
    ejerciciosId?: Array<string>;
    /**
     * 
     * @type {Array<DTOMaterial>}
     * @memberof DTOClaseInscrita
     */
    material?: Array<DTOMaterial>;
    /**
     * 
     * @type {string}
     * @memberof DTOClaseInscrita
     */
    tipoClase?: string;
    /**
     * 
     * @type {DTOProfesor}
     * @memberof DTOClaseInscrita
     */
    profesor?: DTOProfesor;
    /**
     * 
     * @type {Date}
     * @memberof DTOClaseInscrita
     */
    fechaInscripcion?: Date;
    /**
     * 
     * @type {number}
     * @memberof DTOClaseInscrita
     */
    numeroProfesores?: number;
    /**
     * 
     * @type {number}
     * @memberof DTOClaseInscrita
     */
    numeroAlumnos?: number;
}


/**
 * @export
 */
export const DTOClaseInscritaPresencialidadEnum = {
    Online: 'ONLINE',
    Presencial: 'PRESENCIAL',
    Hibrido: 'HIBRIDO'
} as const;
export type DTOClaseInscritaPresencialidadEnum = typeof DTOClaseInscritaPresencialidadEnum[keyof typeof DTOClaseInscritaPresencialidadEnum];

/**
 * @export
 */
export const DTOClaseInscritaNivelEnum = {
    Principiante: 'PRINCIPIANTE',
    Intermedio: 'INTERMEDIO',
    Avanzado: 'AVANZADO',
    Basico: 'BASICO'
} as const;
export type DTOClaseInscritaNivelEnum = typeof DTOClaseInscritaNivelEnum[keyof typeof DTOClaseInscritaNivelEnum];