// frontend/src/lib/generated/api/models/DTOClase.ts
export interface DTOClase {
    /**
     * 
     * @type {number}
     * @memberof DTOClase
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof DTOClase
     */
    titulo?: string;
    /**
     * 
     * @type {string}
     * @memberof DTOClase
     */
    descripcion?: string;
    /**
     * 
     * @type {number}
     * @memberof DTOClase
     */
    precio?: number;
    /**
     * 
     * @type {string}
     * @memberof DTOClase
     */
    presencialidad?: DTOClasePresencialidadEnum;
    /**
     * 
     * @type {string}
     * @memberof DTOClase
     */
    imagenPortada?: string;
    /**
     * 
     * @type {string}
     * @memberof DTOClase
     */
    nivel?: DTOClaseNivelEnum;
    /**
     * 
     * @type {Array<string>}
     * @memberof DTOClase
     */
    alumnosId?: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof DTOClase
     */
    profesoresId?: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof DTOClase
     */
    ejerciciosId?: Array<string>;
    /**
     * 
     * @type {Array<DTOMaterial>}
     * @memberof DTOClase
     */
    material?: Array<DTOMaterial>;
    /**
     * 
     * @type {string}
     * @memberof DTOClase
     */
    tipoClase?: string;
    /**
     * 
     * @type {number}
     * @memberof DTOClase
     */
    numeroProfesores?: number;
    /**
     * 
     * @type {number}
     * @memberof DTOClase
     */
    numeroAlumnos?: number;
    /**
     * 
     * @type {number}
     * @memberof DTOClase
     */
    numeroMateriales?: number;
    /**
     * 
     * @type {number}
     * @memberof DTOClase
     */
    numeroEjercicios?: number;
}


/**
 * @export
 */
export const DTOClasePresencialidadEnum = {
    Online: 'ONLINE',
    Presencial: 'PRESENCIAL',
    Hibrido: 'HIBRIDO'
} as const;
export type DTOClasePresencialidadEnum = typeof DTOClasePresencialidadEnum[keyof typeof DTOClasePresencialidadEnum];

/**
 * @export
 */
export const DTOClaseNivelEnum = {
    Principiante: 'PRINCIPIANTE',
    Intermedio: 'INTERMEDIO',
    Avanzado: 'AVANZADO',
    Basico: 'BASICO'
} as const;
export type DTOClaseNivelEnum = typeof DTOClaseNivelEnum[keyof typeof DTOClaseNivelEnum];